Разбивает код на модули, обозначает файловую струткуру
