+++ js/editor.js
@@ -1,0 +1,1 @@
+/*Здесь будет функциональность по редактированию фотографий*/
