+++ js/filter.js
@@ -1,0 +1,1 @@
+/*Функционал по фильтрации изображений*/
