+++ js/form.js
@@ -1,0 +1,1 @@
+/* Функционал связанный с отправкой данных на сервер */
