+++ js/upload.js
@@ -1,0 +1,1 @@
+/* Здесь будет функциональность по загрузке изображений */
